Revert "initial refactor commit"
